+++ routes/test-routes.js
@@ -8,9 +8,7 @@ var path = require("path");
 // REQUIRE THE MODELS FOLDER
 var db = require("../models");
 var User = require('../models/user.js');
-var passport = require('passport')
-var crypto = require('crypto')
-var jwt = require('jsonwebtoken')
+
 
 var salt = 'asdffdsa'
 // Routes
@@ -47,48 +45,8 @@ module.exports = function (app) {
             res.render("poll", hbsObject);
         });
 
-    //login route
-    app.post("/login" , 
-        passport.authenticate('local', { session: false }),   
-        function (req, res) {
-            res.send({ token: jwt.sign(req.user.dataValues, salt) })
-        }
-    )
+    
 
-    //sign up route
-    app.post("/signup", function (req, res) {
-        db.user.findOne({
-            where: {
-                userName: req.body.userName,
-            }
-        })
-        .then(user => {
-            console.log(user);
-            if(user) {
-                res.status(400).send({ message: 'username is taken, try with different username' })
-            } else {
-                var newUser = db.user.build(req.body)
-                var saltedPassword = newUser.password + salt
-                var hashedPassword = crypto.createHash('md5').update(saltedPassword).digest('hex')
-                // console.log('hashed password', hashedPassword)
-                // console.log('raw password', req.body.password)
-                // console.log('salt' + salt)
-                // console.log('salted password', saltedPassword)
-
-                newUser.password = hashedPassword
-                newUser.save()
-                .then(() => {
-                    res.send({ message: 'success' })
-                })
-                .catch(err => {
-                    res.status(400).send({ message: 'could not sign you up, try again later' })
-                });
-            }
-        })
-        .catch(err => {
-            console.log(err)
-            res.status(400).send({ message: 'username is invalid'})
-        })
     app.get("/popular", function (req, res) {
         var hbsObject = {
             name: 1
@@ -140,21 +98,5 @@ module.exports = function (app) {
             done()
         })
     } 
-
-    //   // cms route loads cms.html
-    //   app.get("/cms", function(req, res) {
-    //     res.sendFile(path.join(__dirname, "../public/cms.html"));
-    //   });
-
-    //   // blog route loads blog.html
-    //   app.get("/blog", function(req, res) {
-    //     res.sendFile(path.join(__dirname, "../public/blog.html"));
-    //   });
-
-    //   // authors route loads author-manager.html
-    //   app.get("/authors", function(req, res) {
-    //     res.sendFile(path.join(__dirname, "../public/author-manager.html"));
-    //   });
-
 });
 }
