+++ server.js
@@ -39,7 +39,8 @@ app.set("view engine", "handlebars");
 
 // Routes
 // =============================================================
-require("./routes/test-routes.js")(app);
+require("./routes/api-routes.js")(app);
+require("./routes/polls-routes.js")(app);
 // require("./routes/author-api-routes.js")(app);
 // require("./routes/post-api-routes.js")(app);
